Add placeholder for how the API examples will live
Our API documentation is being tool generated from our docblock comments in HHVM.
Each function should have an associated example, but we do not want to put the example
in the docblock comment. So the examples can live in the user doc repo and we can merge
the actual API definition and documentation with each example from here.