Remove numbering prefix from all markdown content
The directories (e.g., `01-types`) and the actual markdown files
(e.g., `01-intro.md`) are good for table of contents ordering when
rendered, etc. But using them in actual references from within
markdown are less useful. If we want to reorder the layout, etc.,
we now only need to change the directory and filenames, and not
all references in the markdown. The tooling will take care of figuring
out where the files actually live, etc.