Wait for render processes to finish
This aims to fix a problem I introduced in 6f6fb00b634402.

APIHTMLBuildStep now still opens up as many processes as the CPU can handle
but waits for them to finish before we proceed to create the index.

GuidesHTMLBuildStep on the other hand do some processing with rendered files
right after rendering them. So keep that build step synchronous.
APIHTMLBuildStep is the one that takes most of the time anyway.