Fork 5 threads for rendering API
The previous approach was too consuming and this is much more safe and
configurable.

We currently don't parse command line arguments, which could change number of
threads (and log level by the way). Let's do this sometimes later.

A downside to this is that it's more bulky in the code :(