Moving directories around to a more learner friendly order.
Sorry for how messy this, but it's kind of unavoidable when the folder names control order. I'm moving this to an order that I think works a lot better to split things into 'quickstart', 'basics', and 'advanced' groupings without yet having the ability to add those groupings. I moved getting started to 01 as I want to add 00 as an "Overview" section that gives a really high level introduction to the features.