Trimming the examples in the Async overview
I've removed any specifics from these examples. Important to note that these overviews are intended for absolute beginners, hence the dumbed-down English. We want to be less concerned with the technical details in these overviews as with selling the 'why use this' of each feature - that said, technical inaccuracies are obviously to be removed!