Fix bugs in heading ID generation
Previously, Markdown like this:

### Foo
### Bar
### Qux

Would output this HTML for the third heading:

<h3 id="foo__bar__qux">Qux</h3>

This isn't desirable because it means that every heading ID will
change if the order is changed or if a heading is added. We now only
consider H2's (not H3 siblings).

Also, previously the following Markdown:

## First H2

### First H3

## Second H2

### Second H3

Would output this HTML for the H3 headings:

<h3 id="second-h2__first-h3">First H3</h3>
<h3 id="second-h2__second-h3">Second H3</h3>

because the last H2 that was found was used, even when it appeared
later in the document (i.e. it's not related to the current H3).

The output will now be:

<h2 id="first-h2">First H2</h2>
<h3 id="first-h2__first-h3">First H3</h3>

<h2 id="second-h2">Second H2</h2>
<h3 id="second-h2__second-h3">Second H3</h3>