Fix media query width changes
Since these CSS files are separate, there's no garuntee about the order
of the rules. The override for `.mainTitle .widthWrapper` from `1000px`
to `1200px` in particular was coming in the wrong order, and since the
two rules had the same precedence, the results came out wrong since they
ended up in the wrong order in the ultimate CSS package file.

Instead, pull some stuff out into variables so that it isn't too
horribly offensive to have the rules and the media query overrides in
the same file, so the order ends up right.