Provide PHP ini settings as support data
A script that scrapes the PHP INI settings page and outputs
each setting to a supporting data file for use in helping
determine which settings HHVM supports.

Couldn't use `ini_get_all()` with a PHP executable because
you need every extension loaded for it to print out all the
ini settings.

Ref #216