Add high search score multiplier if a term matches a full word
Was currently only being handled as a full substring match, which has an
extreme penalty.

Additionally, search results do not return anything with a score < 0.3 *
max-score-for-this-search.

This had the end result that if there were any prefix matches (e.g.
'FooBar' in API), a search for Foo would only show them, not guide pages
with 'Foo' in the title.

fixes #658