Use hh-apidoc library
hh-apidoc was split out from this website - but after that refactoring,
hh-apidoc was never actually removed from here. Switch to using the
library.

This reduces duplication, and the library supports documenting type
aliases while the old copy doesn't.

For now, skip documenting type aliases, and add that thoroughly in a
follow-up diff, so that this change is hopefully a no-op.