Update release schedule documentation
Overall content was accurate, but was describing a future change. That
change happened before all future releases.

Also, remove the majority of the supported release table; perhaps bring
back and generate it if we start automating updates for docs.hhvm.com,
but for now, it's stale more often than not and it's weekly manual
maintenance work.

refs facebook/hhvm#8619