Remove discussion of open and closed generic types
Open/closed generics is a C# concept that doesn't exist in Hack. It's
useful to talk about type inference of generics, and erasure, but
open/closed is not a concept in Hack. It also risks confusion with
open shapes.

The main piece of information on this page is discussing that erased
generics cannot be used on static properties, which is already covered
on the erasure page.

Fixes #794