don't deploy in cron runs
AWS doesn't always handle these frequent deployments well (I had to restart the instances a few times because they ran out of memory), so let's skip that part.

We still test the HHVM repomode build, but we won't be able to detect issues with the EB script beyond `eb init` and `eb status` (hopefully that's enough).