add hackttp patch (#920)
Updating the docs site to the latest HHVM version often gets temporarily blocked because the [hackttp](https://github.com/usox/hackttp) dependency causes Hack errors with the respective HHVM version (see https://github.com/usox/hackttp/commits/master -- pretty much every recent commit is due to that). The owner has been pretty responsive about merging our PRs and tagging releases, but it does occasionally take a while so it would be good if we didn't have to rely on it.

This adds the option to automatically patch the hackttp package after composer install/update, with an arbitrary patch (the included patch is from the latest hackttp commit, which didn't make it into a tagged release yet). We can revert this when a new hackttp release is tagged, then re-introduce it whenever we need it again.