Document basics of encoding/locale Str\
I've not documented creating/modifying locales, as that's really
documenting locales, not string behaviors; that said, I don't know
where the best place to document locales is. Maybe it's here; for now
I've aimed to document just enough of `Locale\` for people to get
started with HSL strings.

While I'm here
- remove `__toString()` example and docs as `__toString()` is deprecated
- remove definition of 'numeric string'; there isn't a consistently used
  definition
- reword

This is a bit less precise, but I think the precision here took away
from readability for new users.