Rework devcontainer to use proxygen server
This means that `hhvm` is the actual docker command; the bad news is that if
hhvm crashes, the container goes down.

I'm doing this anyway as there doesn't appear to be support in
devcontainer.json for a separate long-running command - only via docker
compose, which:

- feels like overkill
- appears not to support devcontainer variables like
  containerWorkspaceFolder; this means that it can't work with GitHub
  codespaces, as the `workspaceMount` devcontainer.json option is not
  supported on codespaces

The lack of support for `workspaceMount` is also why there's the fun
messing around with symlinking /var/www and configuration files.