Remove 'banning lval as an expression' page
Generally we don't want to repeat all the type checker rules in the
docs site: they're often complex and it's easy for the site to get out
of sync.

This particular page is old (e.g. it still discusses reference syntax)
and refers to finished work in the future tense, which is confusing.