Rework SASS, remove all Ruby dependencies
This is a little clunky, but given the build's completely broken at the
moment, I'd rather ship this first, and we can revisit it later if
there's a need.

We need to unblock other contributions asap.

The primary goal is to remove the dependency on Ruby SASS, as the
version we were using is no longer installable. Switch to Dart SASS.

This led to:
- stop using 'transition()` mixing from Bourbon:
  - newer bourbons don't have it
  - pretty much every browser supports it nowadays
  - if they don't... we've lost a transition, which isn't a huge deal
- ... and now we don't need Bourbon
- Remove stale references to the Hack font in the CSS
- Upgrade from font-awesome 4.x to 5.x as I couldn't find a good
  non-Rubygems source for 4.5